[Gold IV] Title: 트리의 지름, Time: 4 ms, Memory: 3528 KB -BaekjoonHub

+++ 백준/Gold/1967. 트리의 지름/트리의 지름.cc
@@ -28,7 +28,7 @@ struct Node {
 		
 		if (parent == nullptr)
 			return;
-		if (count == children.size())
+		if (count == (int)children.size())
 			parent->plus(t);
 
 		return;
